fixes enter button issue and separates out auth
